Support user token instead of service token
This adds support for user tokens, giving the user a dropdown
for all of their available graph+variant combinations.

Unfortunately, this list is unfiltered, meaning not just federated
graphs show up. In order to filter this list we'd need to overfetch
or adjust the schema to our needs. For now, nothing happens if you
select a non-federated service.